requirements.txt
Flask
flask-cors
python-dotenv
sendgrid
gunicorn